Added more tests. All tests now pass.
New configuration options. Will be upgrading to the latest minor revision.
There will be an intermediate version on the 1.x branch that will deprecate the
current methods for accessing API results. 2.x will have a new request system,
implemented here on the `rewrite' branch. Errors will be more comprehensive,
and the flow should be much better from one call to the next in the new system.